refactor: extract web.js from main.js (BAT-196)
Move ~340 lines of HTTP/web utilities into dedicated web.js module:
- httpRequest(), web cache (cacheGet/cacheSet), htmlToMarkdown()
- searchBrave(), searchPerplexity(), searchDDG(), searchDDGLite()
- webFetch() with redirect handling + SSRF protection

Depends on config.js for config, log, USER_AGENT. No logic changes.
